@@ -1,18 +1,17 @@
-module github.com/devfullcycle/14-gRPC
+module github.com/GabrieIIeaI/14-gRPC
 
-go 1.19
+go 1.23.2
 
 require (
-	github.com/google/uuid v1.3.0
-	github.com/mattn/go-sqlite3 v1.14.16
-	google.golang.org/grpc v1.50.1
-	google.golang.org/protobuf v1.28.1
+	github.com/google/uuid v1.6.0
+	github.com/mattn/go-sqlite3 v1.14.24
+	google.golang.org/grpc v1.67.1
+	google.golang.org/protobuf v1.35.1
 )
 
 require (
-	github.com/golang/protobuf v1.5.2 // indirect
-	golang.org/x/net v0.0.0-20201021035429-f5854403a974 // indirect
-	golang.org/x/sys v0.0.0-20210119212857-b64e53b001e4 // indirect
-	golang.org/x/text v0.3.3 // indirect
-	google.golang.org/genproto v0.0.0-20200526211855-cb27e3aa2013 // indirect
+	golang.org/x/net v0.28.0 // indirect
+	golang.org/x/sys v0.24.0 // indirect
+	golang.org/x/text v0.17.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20240814211410-ddb44dafa142 // indirect
 )
