@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v3.21.9
+// 	protoc-gen-go v1.35.1
+// 	protoc        v3.6.1
 // source: proto/course_category.proto
 
 package pb
@@ -28,11 +28,9 @@ type Blank struct {
 
 func (x *Blank) Reset() {
 	*x = Blank{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_course_category_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_course_category_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Blank) String() string {
@@ -43,7 +41,7 @@ func (*Blank) ProtoMessage() {}
 
 func (x *Blank) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_course_category_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -70,11 +68,9 @@ type Category struct {
 
 func (x *Category) Reset() {
 	*x = Category{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_course_category_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_course_category_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Category) String() string {
@@ -85,7 +81,7 @@ func (*Category) ProtoMessage() {}
 
 func (x *Category) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_course_category_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -132,11 +128,9 @@ type CreateCategoryRequest struct {
 
 func (x *CreateCategoryRequest) Reset() {
 	*x = CreateCategoryRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_course_category_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_course_category_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CreateCategoryRequest) String() string {
@@ -147,7 +141,7 @@ func (*CreateCategoryRequest) ProtoMessage() {}
 
 func (x *CreateCategoryRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_course_category_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -186,11 +180,9 @@ type CategoryList struct {
 
 func (x *CategoryList) Reset() {
 	*x = CategoryList{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_course_category_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_course_category_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CategoryList) String() string {
@@ -201,7 +193,7 @@ func (*CategoryList) ProtoMessage() {}
 
 func (x *CategoryList) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_course_category_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -223,7 +215,7 @@ func (x *CategoryList) GetCategories() []*Category {
 	return nil
 }
 
-type CategoryGetRequest struct {
+type GetCategoryFilters struct {
 	state         protoimpl.MessageState
 	sizeCache     protoimpl.SizeCache
 	unknownFields protoimpl.UnknownFields
@@ -231,24 +223,22 @@ type CategoryGetRequest struct {
 	Id string `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
 }
 
-func (x *CategoryGetRequest) Reset() {
-	*x = CategoryGetRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_course_category_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+func (x *GetCategoryFilters) Reset() {
+	*x = GetCategoryFilters{}
+	mi := &file_proto_course_category_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
-func (x *CategoryGetRequest) String() string {
+func (x *GetCategoryFilters) String() string {
 	return protoimpl.X.MessageStringOf(x)
 }
 
-func (*CategoryGetRequest) ProtoMessage() {}
+func (*GetCategoryFilters) ProtoMessage() {}
 
-func (x *CategoryGetRequest) ProtoReflect() protoreflect.Message {
+func (x *GetCategoryFilters) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_course_category_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -258,12 +248,12 @@ func (x *CategoryGetRequest) ProtoReflect() protoreflect.Message {
 	return mi.MessageOf(x)
 }
 
-// Deprecated: Use CategoryGetRequest.ProtoReflect.Descriptor instead.
-func (*CategoryGetRequest) Descriptor() ([]byte, []int) {
+// Deprecated: Use GetCategoryFilters.ProtoReflect.Descriptor instead.
+func (*GetCategoryFilters) Descriptor() ([]byte, []int) {
 	return file_proto_course_category_proto_rawDescGZIP(), []int{4}
 }
 
-func (x *CategoryGetRequest) GetId() string {
+func (x *GetCategoryFilters) GetId() string {
 	if x != nil {
 		return x.Id
 	}
@@ -289,10 +279,10 @@ var file_proto_course_category_proto_rawDesc = []byte{
 	0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x4c, 0x69, 0x73, 0x74, 0x12, 0x2c, 0x0a, 0x0a, 0x63,
 	0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x69, 0x65, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x0b, 0x32,
 	0x0c, 0x2e, 0x70, 0x62, 0x2e, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x52, 0x0a, 0x63,
-	0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x69, 0x65, 0x73, 0x22, 0x24, 0x0a, 0x12, 0x43, 0x61, 0x74,
-	0x65, 0x67, 0x6f, 0x72, 0x79, 0x47, 0x65, 0x74, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12,
+	0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x69, 0x65, 0x73, 0x22, 0x24, 0x0a, 0x12, 0x47, 0x65, 0x74,
+	0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x12,
 	0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x64, 0x32,
-	0xd3, 0x02, 0x0a, 0x0f, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x53, 0x65, 0x72, 0x76,
+	0xd7, 0x02, 0x0a, 0x0f, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x53, 0x65, 0x72, 0x76,
 	0x69, 0x63, 0x65, 0x12, 0x3b, 0x0a, 0x0e, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x43, 0x61, 0x74,
 	0x65, 0x67, 0x6f, 0x72, 0x79, 0x12, 0x19, 0x2e, 0x70, 0x62, 0x2e, 0x43, 0x72, 0x65, 0x61, 0x74,
 	0x65, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74,
@@ -309,12 +299,12 @@ var file_proto_course_category_proto_rawDesc = []byte{
 	0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x22, 0x00, 0x28, 0x01, 0x30, 0x01, 0x12, 0x2f, 0x0a,
 	0x0e, 0x4c, 0x69, 0x73, 0x74, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x69, 0x65, 0x73, 0x12,
 	0x09, 0x2e, 0x70, 0x62, 0x2e, 0x62, 0x6c, 0x61, 0x6e, 0x6b, 0x1a, 0x10, 0x2e, 0x70, 0x62, 0x2e,
-	0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x4c, 0x69, 0x73, 0x74, 0x22, 0x00, 0x12, 0x35,
-	0x0a, 0x0b, 0x47, 0x65, 0x74, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x12, 0x16, 0x2e,
-	0x70, 0x62, 0x2e, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x47, 0x65, 0x74, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x0c, 0x2e, 0x70, 0x62, 0x2e, 0x43, 0x61, 0x74, 0x65, 0x67,
-	0x6f, 0x72, 0x79, 0x22, 0x00, 0x42, 0x0d, 0x5a, 0x0b, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61,
-	0x6c, 0x2f, 0x70, 0x62, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+	0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x4c, 0x69, 0x73, 0x74, 0x22, 0x00, 0x12, 0x39,
+	0x0a, 0x0f, 0x47, 0x65, 0x74, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x42, 0x79, 0x49,
+	0x44, 0x12, 0x16, 0x2e, 0x70, 0x62, 0x2e, 0x47, 0x65, 0x74, 0x43, 0x61, 0x74, 0x65, 0x67, 0x6f,
+	0x72, 0x79, 0x46, 0x69, 0x6c, 0x74, 0x65, 0x72, 0x73, 0x1a, 0x0c, 0x2e, 0x70, 0x62, 0x2e, 0x43,
+	0x61, 0x74, 0x65, 0x67, 0x6f, 0x72, 0x79, 0x22, 0x00, 0x42, 0x0d, 0x5a, 0x0b, 0x69, 0x6e, 0x74,
+	0x65, 0x72, 0x6e, 0x61, 0x6c, 0x2f, 0x70, 0x62, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
 }
 
 var (
@@ -330,12 +320,12 @@ func file_proto_course_category_proto_rawDescGZIP() []byte {
 }
 
 var file_proto_course_category_proto_msgTypes = make([]protoimpl.MessageInfo, 5)
-var file_proto_course_category_proto_goTypes = []interface{}{
+var file_proto_course_category_proto_goTypes = []any{
 	(*Blank)(nil),                 // 0: pb.blank
 	(*Category)(nil),              // 1: pb.Category
 	(*CreateCategoryRequest)(nil), // 2: pb.CreateCategoryRequest
 	(*CategoryList)(nil),          // 3: pb.CategoryList
-	(*CategoryGetRequest)(nil),    // 4: pb.CategoryGetRequest
+	(*GetCategoryFilters)(nil),    // 4: pb.GetCategoryFilters
 }
 var file_proto_course_category_proto_depIdxs = []int32{
 	1, // 0: pb.CategoryList.categories:type_name -> pb.Category
@@ -343,12 +333,12 @@ var file_proto_course_category_proto_depIdxs = []int32{
 	2, // 2: pb.CategoryService.CreateCategoryStream:input_type -> pb.CreateCategoryRequest
 	2, // 3: pb.CategoryService.CreateCategoryStreamBidirectional:input_type -> pb.CreateCategoryRequest
 	0, // 4: pb.CategoryService.ListCategories:input_type -> pb.blank
-	4, // 5: pb.CategoryService.GetCategory:input_type -> pb.CategoryGetRequest
+	4, // 5: pb.CategoryService.GetCategoryByID:input_type -> pb.GetCategoryFilters
 	1, // 6: pb.CategoryService.CreateCategory:output_type -> pb.Category
 	3, // 7: pb.CategoryService.CreateCategoryStream:output_type -> pb.CategoryList
 	1, // 8: pb.CategoryService.CreateCategoryStreamBidirectional:output_type -> pb.Category
 	3, // 9: pb.CategoryService.ListCategories:output_type -> pb.CategoryList
-	1, // 10: pb.CategoryService.GetCategory:output_type -> pb.Category
+	1, // 10: pb.CategoryService.GetCategoryByID:output_type -> pb.Category
 	6, // [6:11] is the sub-list for method output_type
 	1, // [1:6] is the sub-list for method input_type
 	1, // [1:1] is the sub-list for extension type_name
@@ -360,68 +350,6 @@ func init() { file_proto_course_category_proto_init() }
 func file_proto_course_category_proto_init() {
 	if File_proto_course_category_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_proto_course_category_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Blank); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_course_category_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Category); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_course_category_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CreateCategoryRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_course_category_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CategoryList); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_course_category_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CategoryGetRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
