@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
-// - protoc             v3.21.9
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v3.6.1
 // source: proto/course_category.proto
 
 package pb
@@ -15,18 +15,26 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
+
+const (
+	CategoryService_CreateCategory_FullMethodName                    = "/pb.CategoryService/CreateCategory"
+	CategoryService_CreateCategoryStream_FullMethodName              = "/pb.CategoryService/CreateCategoryStream"
+	CategoryService_CreateCategoryStreamBidirectional_FullMethodName = "/pb.CategoryService/CreateCategoryStreamBidirectional"
+	CategoryService_ListCategories_FullMethodName                    = "/pb.CategoryService/ListCategories"
+	CategoryService_GetCategoryByID_FullMethodName                   = "/pb.CategoryService/GetCategoryByID"
+)
 
 // CategoryServiceClient is the client API for CategoryService service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
 type CategoryServiceClient interface {
 	CreateCategory(ctx context.Context, in *CreateCategoryRequest, opts ...grpc.CallOption) (*Category, error)
-	CreateCategoryStream(ctx context.Context, opts ...grpc.CallOption) (CategoryService_CreateCategoryStreamClient, error)
-	CreateCategoryStreamBidirectional(ctx context.Context, opts ...grpc.CallOption) (CategoryService_CreateCategoryStreamBidirectionalClient, error)
+	CreateCategoryStream(ctx context.Context, opts ...grpc.CallOption) (grpc.ClientStreamingClient[CreateCategoryRequest, CategoryList], error)
+	CreateCategoryStreamBidirectional(ctx context.Context, opts ...grpc.CallOption) (grpc.BidiStreamingClient[CreateCategoryRequest, Category], error)
 	ListCategories(ctx context.Context, in *Blank, opts ...grpc.CallOption) (*CategoryList, error)
-	GetCategory(ctx context.Context, in *CategoryGetRequest, opts ...grpc.CallOption) (*Category, error)
+	GetCategoryByID(ctx context.Context, in *GetCategoryFilters, opts ...grpc.CallOption) (*Category, error)
 }
 
 type categoryServiceClient struct {
@@ -38,91 +46,55 @@ func NewCategoryServiceClient(cc grpc.ClientConnInterface) CategoryServiceClient
 }
 
 func (c *categoryServiceClient) CreateCategory(ctx context.Context, in *CreateCategoryRequest, opts ...grpc.CallOption) (*Category, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Category)
-	err := c.cc.Invoke(ctx, "/pb.CategoryService/CreateCategory", in, out, opts...)
+	err := c.cc.Invoke(ctx, CategoryService_CreateCategory_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *categoryServiceClient) CreateCategoryStream(ctx context.Context, opts ...grpc.CallOption) (CategoryService_CreateCategoryStreamClient, error) {
-	stream, err := c.cc.NewStream(ctx, &CategoryService_ServiceDesc.Streams[0], "/pb.CategoryService/CreateCategoryStream", opts...)
+func (c *categoryServiceClient) CreateCategoryStream(ctx context.Context, opts ...grpc.CallOption) (grpc.ClientStreamingClient[CreateCategoryRequest, CategoryList], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &CategoryService_ServiceDesc.Streams[0], CategoryService_CreateCategoryStream_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &categoryServiceCreateCategoryStreamClient{stream}
+	x := &grpc.GenericClientStream[CreateCategoryRequest, CategoryList]{ClientStream: stream}
 	return x, nil
 }
 
-type CategoryService_CreateCategoryStreamClient interface {
-	Send(*CreateCategoryRequest) error
-	CloseAndRecv() (*CategoryList, error)
-	grpc.ClientStream
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CategoryService_CreateCategoryStreamClient = grpc.ClientStreamingClient[CreateCategoryRequest, CategoryList]
 
-type categoryServiceCreateCategoryStreamClient struct {
-	grpc.ClientStream
-}
-
-func (x *categoryServiceCreateCategoryStreamClient) Send(m *CreateCategoryRequest) error {
-	return x.ClientStream.SendMsg(m)
-}
-
-func (x *categoryServiceCreateCategoryStreamClient) CloseAndRecv() (*CategoryList, error) {
-	if err := x.ClientStream.CloseSend(); err != nil {
-		return nil, err
-	}
-	m := new(CategoryList)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
-
-func (c *categoryServiceClient) CreateCategoryStreamBidirectional(ctx context.Context, opts ...grpc.CallOption) (CategoryService_CreateCategoryStreamBidirectionalClient, error) {
-	stream, err := c.cc.NewStream(ctx, &CategoryService_ServiceDesc.Streams[1], "/pb.CategoryService/CreateCategoryStreamBidirectional", opts...)
+func (c *categoryServiceClient) CreateCategoryStreamBidirectional(ctx context.Context, opts ...grpc.CallOption) (grpc.BidiStreamingClient[CreateCategoryRequest, Category], error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	stream, err := c.cc.NewStream(ctx, &CategoryService_ServiceDesc.Streams[1], CategoryService_CreateCategoryStreamBidirectional_FullMethodName, cOpts...)
 	if err != nil {
 		return nil, err
 	}
-	x := &categoryServiceCreateCategoryStreamBidirectionalClient{stream}
+	x := &grpc.GenericClientStream[CreateCategoryRequest, Category]{ClientStream: stream}
 	return x, nil
 }
 
-type CategoryService_CreateCategoryStreamBidirectionalClient interface {
-	Send(*CreateCategoryRequest) error
-	Recv() (*Category, error)
-	grpc.ClientStream
-}
-
-type categoryServiceCreateCategoryStreamBidirectionalClient struct {
-	grpc.ClientStream
-}
-
-func (x *categoryServiceCreateCategoryStreamBidirectionalClient) Send(m *CreateCategoryRequest) error {
-	return x.ClientStream.SendMsg(m)
-}
-
-func (x *categoryServiceCreateCategoryStreamBidirectionalClient) Recv() (*Category, error) {
-	m := new(Category)
-	if err := x.ClientStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CategoryService_CreateCategoryStreamBidirectionalClient = grpc.BidiStreamingClient[CreateCategoryRequest, Category]
 
 func (c *categoryServiceClient) ListCategories(ctx context.Context, in *Blank, opts ...grpc.CallOption) (*CategoryList, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CategoryList)
-	err := c.cc.Invoke(ctx, "/pb.CategoryService/ListCategories", in, out, opts...)
+	err := c.cc.Invoke(ctx, CategoryService_ListCategories_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
 	return out, nil
 }
 
-func (c *categoryServiceClient) GetCategory(ctx context.Context, in *CategoryGetRequest, opts ...grpc.CallOption) (*Category, error) {
+func (c *categoryServiceClient) GetCategoryByID(ctx context.Context, in *GetCategoryFilters, opts ...grpc.CallOption) (*Category, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Category)
-	err := c.cc.Invoke(ctx, "/pb.CategoryService/GetCategory", in, out, opts...)
+	err := c.cc.Invoke(ctx, CategoryService_GetCategoryByID_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -131,36 +103,40 @@ func (c *categoryServiceClient) GetCategory(ctx context.Context, in *CategoryGet
 
 // CategoryServiceServer is the server API for CategoryService service.
 // All implementations must embed UnimplementedCategoryServiceServer
-// for forward compatibility
+// for forward compatibility.
 type CategoryServiceServer interface {
 	CreateCategory(context.Context, *CreateCategoryRequest) (*Category, error)
-	CreateCategoryStream(CategoryService_CreateCategoryStreamServer) error
-	CreateCategoryStreamBidirectional(CategoryService_CreateCategoryStreamBidirectionalServer) error
+	CreateCategoryStream(grpc.ClientStreamingServer[CreateCategoryRequest, CategoryList]) error
+	CreateCategoryStreamBidirectional(grpc.BidiStreamingServer[CreateCategoryRequest, Category]) error
 	ListCategories(context.Context, *Blank) (*CategoryList, error)
-	GetCategory(context.Context, *CategoryGetRequest) (*Category, error)
+	GetCategoryByID(context.Context, *GetCategoryFilters) (*Category, error)
 	mustEmbedUnimplementedCategoryServiceServer()
 }
 
-// UnimplementedCategoryServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedCategoryServiceServer struct {
-}
+// UnimplementedCategoryServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedCategoryServiceServer struct{}
 
 func (UnimplementedCategoryServiceServer) CreateCategory(context.Context, *CreateCategoryRequest) (*Category, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method CreateCategory not implemented")
 }
-func (UnimplementedCategoryServiceServer) CreateCategoryStream(CategoryService_CreateCategoryStreamServer) error {
+func (UnimplementedCategoryServiceServer) CreateCategoryStream(grpc.ClientStreamingServer[CreateCategoryRequest, CategoryList]) error {
 	return status.Errorf(codes.Unimplemented, "method CreateCategoryStream not implemented")
 }
-func (UnimplementedCategoryServiceServer) CreateCategoryStreamBidirectional(CategoryService_CreateCategoryStreamBidirectionalServer) error {
+func (UnimplementedCategoryServiceServer) CreateCategoryStreamBidirectional(grpc.BidiStreamingServer[CreateCategoryRequest, Category]) error {
 	return status.Errorf(codes.Unimplemented, "method CreateCategoryStreamBidirectional not implemented")
 }
 func (UnimplementedCategoryServiceServer) ListCategories(context.Context, *Blank) (*CategoryList, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListCategories not implemented")
 }
-func (UnimplementedCategoryServiceServer) GetCategory(context.Context, *CategoryGetRequest) (*Category, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method GetCategory not implemented")
+func (UnimplementedCategoryServiceServer) GetCategoryByID(context.Context, *GetCategoryFilters) (*Category, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetCategoryByID not implemented")
 }
 func (UnimplementedCategoryServiceServer) mustEmbedUnimplementedCategoryServiceServer() {}
+func (UnimplementedCategoryServiceServer) testEmbeddedByValue()                         {}
 
 // UnsafeCategoryServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to CategoryServiceServer will
@@ -170,6 +146,13 @@ type UnsafeCategoryServiceServer interface {
 }
 
 func RegisterCategoryServiceServer(s grpc.ServiceRegistrar, srv CategoryServiceServer) {
+	// If the following call pancis, it indicates UnimplementedCategoryServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&CategoryService_ServiceDesc, srv)
 }
 
@@ -183,7 +166,7 @@ func _CategoryService_CreateCategory_Handler(srv interface{}, ctx context.Contex
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pb.CategoryService/CreateCategory",
+		FullMethod: CategoryService_CreateCategory_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(CategoryServiceServer).CreateCategory(ctx, req.(*CreateCategoryRequest))
@@ -192,56 +175,18 @@ func _CategoryService_CreateCategory_Handler(srv interface{}, ctx context.Contex
 }
 
 func _CategoryService_CreateCategoryStream_Handler(srv interface{}, stream grpc.ServerStream) error {
-	return srv.(CategoryServiceServer).CreateCategoryStream(&categoryServiceCreateCategoryStreamServer{stream})
+	return srv.(CategoryServiceServer).CreateCategoryStream(&grpc.GenericServerStream[CreateCategoryRequest, CategoryList]{ServerStream: stream})
 }
 
-type CategoryService_CreateCategoryStreamServer interface {
-	SendAndClose(*CategoryList) error
-	Recv() (*CreateCategoryRequest, error)
-	grpc.ServerStream
-}
-
-type categoryServiceCreateCategoryStreamServer struct {
-	grpc.ServerStream
-}
-
-func (x *categoryServiceCreateCategoryStreamServer) SendAndClose(m *CategoryList) error {
-	return x.ServerStream.SendMsg(m)
-}
-
-func (x *categoryServiceCreateCategoryStreamServer) Recv() (*CreateCategoryRequest, error) {
-	m := new(CreateCategoryRequest)
-	if err := x.ServerStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CategoryService_CreateCategoryStreamServer = grpc.ClientStreamingServer[CreateCategoryRequest, CategoryList]
 
 func _CategoryService_CreateCategoryStreamBidirectional_Handler(srv interface{}, stream grpc.ServerStream) error {
-	return srv.(CategoryServiceServer).CreateCategoryStreamBidirectional(&categoryServiceCreateCategoryStreamBidirectionalServer{stream})
+	return srv.(CategoryServiceServer).CreateCategoryStreamBidirectional(&grpc.GenericServerStream[CreateCategoryRequest, Category]{ServerStream: stream})
 }
 
-type CategoryService_CreateCategoryStreamBidirectionalServer interface {
-	Send(*Category) error
-	Recv() (*CreateCategoryRequest, error)
-	grpc.ServerStream
-}
-
-type categoryServiceCreateCategoryStreamBidirectionalServer struct {
-	grpc.ServerStream
-}
-
-func (x *categoryServiceCreateCategoryStreamBidirectionalServer) Send(m *Category) error {
-	return x.ServerStream.SendMsg(m)
-}
-
-func (x *categoryServiceCreateCategoryStreamBidirectionalServer) Recv() (*CreateCategoryRequest, error) {
-	m := new(CreateCategoryRequest)
-	if err := x.ServerStream.RecvMsg(m); err != nil {
-		return nil, err
-	}
-	return m, nil
-}
+// This type alias is provided for backwards compatibility with existing code that references the prior non-generic stream type by name.
+type CategoryService_CreateCategoryStreamBidirectionalServer = grpc.BidiStreamingServer[CreateCategoryRequest, Category]
 
 func _CategoryService_ListCategories_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(Blank)
@@ -253,7 +198,7 @@ func _CategoryService_ListCategories_Handler(srv interface{}, ctx context.Contex
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pb.CategoryService/ListCategories",
+		FullMethod: CategoryService_ListCategories_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(CategoryServiceServer).ListCategories(ctx, req.(*Blank))
@@ -261,20 +206,20 @@ func _CategoryService_ListCategories_Handler(srv interface{}, ctx context.Contex
 	return interceptor(ctx, in, info, handler)
 }
 
-func _CategoryService_GetCategory_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(CategoryGetRequest)
+func _CategoryService_GetCategoryByID_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(GetCategoryFilters)
 	if err := dec(in); err != nil {
 		return nil, err
 	}
 	if interceptor == nil {
-		return srv.(CategoryServiceServer).GetCategory(ctx, in)
+		return srv.(CategoryServiceServer).GetCategoryByID(ctx, in)
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/pb.CategoryService/GetCategory",
+		FullMethod: CategoryService_GetCategoryByID_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(CategoryServiceServer).GetCategory(ctx, req.(*CategoryGetRequest))
+		return srv.(CategoryServiceServer).GetCategoryByID(ctx, req.(*GetCategoryFilters))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -295,8 +240,8 @@ var CategoryService_ServiceDesc = grpc.ServiceDesc{
 			Handler:    _CategoryService_ListCategories_Handler,
 		},
 		{
-			MethodName: "GetCategory",
-			Handler:    _CategoryService_GetCategory_Handler,
+			MethodName: "GetCategoryByID",
+			Handler:    _CategoryService_GetCategoryByID_Handler,
 		},
 	},
 	Streams: []grpc.StreamDesc{
