@@ -4,8 +4,8 @@ import (
 	"context"
 	"io"
 
-	"github.com/devfullcycle/14-gRPC/internal/database"
-	"github.com/devfullcycle/14-gRPC/internal/pb"
+	"github.com/GabrieIIeaI/14-gRPC/internal/database"
+	"github.com/GabrieIIeaI/14-gRPC/internal/pb"
 )
 
 type CategoryService struct {
@@ -21,13 +21,14 @@ func NewCategoryService(categoryDB database.Category) *CategoryService {
 
 func (c *CategoryService) CreateCategory(ctx context.Context, in *pb.CreateCategoryRequest) (*pb.Category, error) {
 	category, err := c.CategoryDB.Create(in.Name, in.Description)
+
 	if err != nil {
 		return nil, err
 	}
 
 	categoryResponse := &pb.Category{
-		Id:          category.ID,
-		Name:        category.Name,
+		Id: category.ID,
+		Name: category.Name,
 		Description: category.Description,
 	}
 
@@ -36,36 +37,38 @@ func (c *CategoryService) CreateCategory(ctx context.Context, in *pb.CreateCateg
 
 func (c *CategoryService) ListCategories(ctx context.Context, in *pb.Blank) (*pb.CategoryList, error) {
 	categories, err := c.CategoryDB.FindAll()
+
 	if err != nil {
 		return nil, err
 	}
 
-	var categoriesResponse []*pb.Category
+	var categoriesResponse = []*pb.Category{}
 
 	for _, category := range categories {
 		categoryResponse := &pb.Category{
-			Id:          category.ID,
-			Name:        category.Name,
+			Id: category.ID,
+			Name: category.Name,
 			Description: category.Description,
 		}
 
 		categoriesResponse = append(categoriesResponse, categoryResponse)
-	}
+	} 
 
 	return &pb.CategoryList{Categories: categoriesResponse}, nil
 }
 
-func (c *CategoryService) GetCategory(ctx context.Context, in *pb.CategoryGetRequest) (*pb.Category, error) {
+func (c *CategoryService) GetCategoryByID(ctx context.Context, in *pb.GetCategoryFilters) (*pb.Category, error) {
 	category, err := c.CategoryDB.Find(in.Id)
+
 	if err != nil {
 		return nil, err
 	}
 
 	categoryResponse := &pb.Category{
-		Id:          category.ID,
-		Name:        category.Name,
+		Id: category.ID,
+		Name: category.Name,
 		Description: category.Description,
-	}
+	} 
 
 	return categoryResponse, nil
 }
@@ -78,6 +81,7 @@ func (c *CategoryService) CreateCategoryStream(stream pb.CategoryService_CreateC
 		if err == io.EOF {
 			return stream.SendAndClose(categories)
 		}
+
 		if err != nil {
 			return err
 		}
@@ -88,8 +92,8 @@ func (c *CategoryService) CreateCategoryStream(stream pb.CategoryService_CreateC
 		}
 
 		categories.Categories = append(categories.Categories, &pb.Category{
-			Id:          categoryResult.ID,
-			Name:        categoryResult.Name,
+			Id: categoryResult.ID,
+			Name: categoryResult.Name,
 			Description: categoryResult.Description,
 		})
 	}
@@ -99,7 +103,7 @@ func (c *CategoryService) CreateCategoryStreamBidirectional(stream pb.CategorySe
 	for {
 		category, err := stream.Recv()
 		if err == io.EOF {
-			return nil
+			return nil 
 		}
 		if err != nil {
 			return err
@@ -110,11 +114,13 @@ func (c *CategoryService) CreateCategoryStreamBidirectional(stream pb.CategorySe
 			return err
 		}
 
-		err = stream.Send(&pb.Category{
-			Id:          categoryResult.ID,
-			Name:        categoryResult.Name,
+		var categoryResponse = &pb.Category{
+			Id: categoryResult.ID,
+			Name: categoryResult.Name,
 			Description: categoryResult.Description,
-		})
+		}
+
+		err = stream.Send(categoryResponse)
 		if err != nil {
 			return err
 		}
